fix: metric service light refactoring
