+++ server/src/main/java/com/hedera/block/server/persistence/storage/write/BlockAsDirWriter.java
@@ -17,7 +17,7 @@
 package com.hedera.block.server.persistence.storage.write;
 
 import static com.hedera.block.server.Constants.BLOCK_FILE_EXTENSION;
-import static com.hedera.block.server.metrics.BlockNodeMetricNames.Counter.BlocksPersisted;
+import static com.hedera.block.server.metrics.BlockNodeMetricTypes.Counter.BlocksPersisted;
 import static java.lang.System.Logger;
 import static java.lang.System.Logger.Level.DEBUG;
 import static java.lang.System.Logger.Level.ERROR;
